refactor(algolia): remove old explorer indexing code (#3472)
