+++ Makefile
@@ -357,7 +357,6 @@ reindex: itsJustJavascript
 	node --enable-source-maps itsJustJavascript/baker/algolia/configureAlgolia.js
 	node --enable-source-maps itsJustJavascript/baker/algolia/indexToAlgolia.js
 	node --enable-source-maps itsJustJavascript/baker/algolia/indexChartsToAlgolia.js
-	node --enable-source-maps itsJustJavascript/baker/algolia/indexExplorersToAlgolia.js
 	node --enable-source-maps itsJustJavascript/baker/algolia/indexExplorerViewsToAlgolia.js
 
 clean:
